+++ apps/workflow-server/gcc-workflow-server-facade/gcc-restclient-facade-default/src/test/java/com/coremedia/labs/translation/gcc/facade/def/DefaultGCExchangeFacadeTest.java
@@ -182,9 +182,12 @@ class DefaultGCExchangeFacadeTest {
     @DisplayName("Test for successful submission.")
     void happyPath(TestInfo testInfo) {
       String subject = testInfo.getDisplayName();
+      String comment = "Test";
       LocalDateTime dueDateLocal = LocalDateTime.now().plusHours(2);
       ZoneOffset zoneOffset = ZoneOffset.ofHoursMinutesSeconds(1, 2, 3);
       ZonedDateTime dueDate = ZonedDateTime.of(dueDateLocal, zoneOffset);
+      String workflow = "pseudo translation";
+      String submitter = "admin";
       Date expectedUtcDueDate = Date.from(dueDateLocal.atOffset(zoneOffset).toInstant());
       Locale sourceLocale = Locale.US;
       Locale targetLocale = Locale.FRANCE;
@@ -196,7 +199,7 @@ class DefaultGCExchangeFacadeTest {
       when(response.getSubmissionId()).thenReturn(expectedSubmissionId);
 
       try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-        long submissionId = facade.submitSubmission(subject, dueDate, sourceLocale, singletonMap(fileId, singletonList(targetLocale)));
+        long submissionId = facade.submitSubmission(subject, comment, dueDate, workflow, submitter, sourceLocale, singletonMap(fileId, singletonList(targetLocale)));
 
         assertThat(submissionId).isEqualTo(expectedSubmissionId);
         verify(gcExchange).submitSubmission(submissionSubmitRequestCaptor.capture());
@@ -215,9 +218,12 @@ class DefaultGCExchangeFacadeTest {
     @DisplayName("Correctly deal with communication errors.")
     void dealWithCommunicationExceptions(TestInfo testInfo) {
       String subject = testInfo.getDisplayName();
+      String comment = "Test";
       LocalDateTime dueDateLocal = LocalDateTime.now().plusHours(2);
       ZoneOffset zoneOffset = ZoneOffset.ofHoursMinutesSeconds(1, 2, 3);
       ZonedDateTime dueDate = ZonedDateTime.of(dueDateLocal, zoneOffset);
+      String workflow = "pseudo translation";
+      String submitter = "admin";
       Locale sourceLocale = Locale.US;
       Locale targetLocale = Locale.FRANCE;
       String fileId = "1234-5678";
@@ -225,7 +231,7 @@ class DefaultGCExchangeFacadeTest {
       when(gcExchange.submitSubmission(any())).thenThrow(RuntimeException.class);
 
       try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-        assertThatThrownBy(() -> facade.submitSubmission(subject, dueDate, sourceLocale, singletonMap(fileId, singletonList(targetLocale))))
+        assertThatThrownBy(() -> facade.submitSubmission(subject, comment, dueDate, workflow, submitter, sourceLocale, singletonMap(fileId, singletonList(targetLocale))))
                 .isInstanceOf(GCFacadeCommunicationException.class)
                 .hasCauseInstanceOf(RuntimeException.class)
                 .hasMessageContaining(subject)
@@ -481,7 +487,7 @@ class DefaultGCExchangeFacadeTest {
     @DisplayName("Standard Submission State Retrieval")
     void happyPath() {
       try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-        GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+        GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
         assertThat(state).isEqualTo(GCSubmissionState.STARTED);
       }
     }
@@ -492,7 +498,7 @@ class DefaultGCExchangeFacadeTest {
       when(submissionsResponseData.getSubmissions()).thenReturn(emptyList());
 
       try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-        GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+        GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
         assertThat(state).isEqualTo(GCSubmissionState.OTHER);
       }
     }
@@ -506,7 +512,7 @@ class DefaultGCExchangeFacadeTest {
       when(submissionsResponseData.getSubmissions()).thenReturn(asList(submission, additionalSubmission));
 
       try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-        GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+        GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
         // Does not matter, just one should be chosen.
         assertThat(state).isIn(GCSubmissionState.STARTED, GCSubmissionState.TRANSLATE);
       }
@@ -540,7 +546,7 @@ class DefaultGCExchangeFacadeTest {
           when(task.getStatus()).thenReturn(TaskStatus.Processing.text());
 
           try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-            GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+            GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
 
             assertThat(state).isEqualTo(GCSubmissionState.CANCELLED);
           }
@@ -553,7 +559,7 @@ class DefaultGCExchangeFacadeTest {
           when(task.getIsCancelConfirmed()).thenReturn(Boolean.FALSE);
 
           try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-            GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+            GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
 
             assertThat(state).isEqualTo(GCSubmissionState.CANCELLED);
           }
@@ -571,7 +577,7 @@ class DefaultGCExchangeFacadeTest {
           when(additionalTask2.getIsCancelConfirmed()).thenReturn(Boolean.TRUE);
 
           try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-            GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+            GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
 
             assertThat(state).isEqualTo(GCSubmissionState.CANCELLED);
           }
@@ -588,7 +594,7 @@ class DefaultGCExchangeFacadeTest {
           when(task.getIsCancelConfirmed()).thenReturn(Boolean.TRUE);
 
           try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-            GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+            GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
 
             assertThat(state).isEqualTo(GCSubmissionState.CANCELLATION_CONFIRMED);
           }
@@ -611,7 +617,7 @@ class DefaultGCExchangeFacadeTest {
           when(additionalTask2.getIsCancelConfirmed()).thenReturn(Boolean.TRUE);
 
           try (GCExchangeFacade facade = new MockDefaultGCExchangeFacade(gcExchange)) {
-            GCSubmissionState state = facade.getSubmissionState(SUBMISSION_ID);
+            GCSubmissionState state = facade.getSubmission(SUBMISSION_ID).getState();
 
             assertThat(state).isEqualTo(GCSubmissionState.CANCELLATION_CONFIRMED);
           }